MCPP/mcp-weather/requirements.txt
langchain
langchain-core
langchain-community
google-generativeai
python-dotenv
langchain-mcp-adapters
langchain-google-genai
mcp
langgraph